Tested with Scala 3 and JDK-21 on Intel® Core™ i7-11800H:

#### Before
```txt
Benchmark                            (size)   Mode  Cnt        Score        Error  Units
AnyValsWriting.zioJson                  N/A  thrpt    5  3823372.648 ±  13023.722  ops/s
AnyValsWriting.zioSchemaJson            N/A  thrpt    5  3852940.951 ± 195576.957  ops/s
ArrayOfDoublesWriting.zioJson           128  thrpt    5   184105.035 ±   3967.936  ops/s
ArrayOfDoublesWriting.zioSchemaJson     128  thrpt    5   161213.205 ±    607.026  ops/s
ArrayOfFloatsWriting.zioJson            128  thrpt    5   244258.544 ±   3256.474  ops/s
ArrayOfFloatsWriting.zioSchemaJson      128  thrpt    5   244916.346 ±   8006.527  ops/s
PrimitivesWriting.zioJson               N/A  thrpt    5  6349196.701 ±  26773.712  ops/s
PrimitivesWriting.zioSchemaJson         N/A  thrpt    5  5217725.547 ±  71947.044  ops/s
```

#### After
```txt
Benchmark                            (size)   Mode  Cnt        Score        Error  Units
AnyValsWriting.zioJson                  N/A  thrpt    5  3850793.395 ± 206318.568  ops/s
AnyValsWriting.zioSchemaJson            N/A  thrpt    5  3927681.798 ±  31279.885  ops/s
ArrayOfDoublesWriting.zioJson           128  thrpt    5   213042.974 ±   1967.557  ops/s
ArrayOfDoublesWriting.zioSchemaJson     128  thrpt    5   210949.814 ±   2905.848  ops/s
ArrayOfFloatsWriting.zioJson            128  thrpt    5   285297.646 ±   7687.284  ops/s
ArrayOfFloatsWriting.zioSchemaJson      128  thrpt    5   314495.765 ±   5640.222  ops/s
PrimitivesWriting.zioJson               N/A  thrpt    5  6801915.370 ± 171719.113  ops/s
PrimitivesWriting.zioSchemaJson         N/A  thrpt    5  5397774.649 ± 354764.552  ops/s
```